добавление простой log.Print -( вместо fmt.Print для консоли)

+++ internal/models/account.go
@@ -1,7 +1,7 @@
 package models
 
 type Account struct {
-	ID      int     `db:"id" json:"id"`
-	Balance float64 `db:"balance" json:"balance"`
-	Owner   string  `db:"owner" json:"owner"`
+	ID      int    `db:"id" json:"id"`
+	Balance int64  `db:"balance" json:"balance"`
+	Owner   string `db:"owner" json:"owner"`
 }
